+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.4.1 - 2022-05-18
 
-- FEAT: withDefault now also sets a new process.env var under `CONFIG_HELPER_` prefix `process.env['CONFIG_HELPER_' + environmentVariable] = process.env[environmentVariable] || defaultValue;`  
+- FEAT: withDefault and required now also sets a new var under `CONFIG_HELPER_` prefix: `process.env['CONFIG_HELPER_' + environmentVariable]`
 
 # 1.3.1 and below - 2020-07-21